Rename mock file trip-event on trip-point.js
